[Natives] Changed Structure
Changed dir/file structure so that each native folder contains the compiled native library inside the sub-folder resources, with the correct platform-dependent path.

This means that maven can include the native libs (with the correct paths) by simply defining the "resources" folder as a maven resource when doing a release.

This makes native copy scripts redundant.