[Natives] Refined Release Process
The build process remains unchanged, but the new "release" profile within natives POM files ensures that native compiling is skipped on release.

Release profile is set up to include the "resources" dir. There is one in each of the native platform directories, containing the compiled native library with the correct path